+++ GAPTest/GAPTest.Web/Controllers/CustomersController.cs
@@ -112,7 +112,6 @@ namespace GAPTest.Web.Controllers
             return View(model);
         }
 
-        // GET: Customers/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
             if (id == null)
@@ -120,47 +119,46 @@ namespace GAPTest.Web.Controllers
                 return NotFound();
             }
 
-            var customer = await _context.Customers.FindAsync(id);
+            var customer = await _context.Customers
+                .Include(o => o.User)
+                .FirstOrDefaultAsync(o => o.Id == id.Value);
             if (customer == null)
             {
                 return NotFound();
             }
-            return View(customer);
+
+            var model = new EditUserViewModel
+            {
+                Address = customer.User.Address,
+                Document = customer.User.Document,
+                Name = customer.User.Name,
+                Id = customer.Id,
+                CellPhone = customer.User.PhoneNumber
+            };
+
+            return View(model);
         }
 
-        // POST: Customers/Edit/5
-        // To protect from overposting attacks, please enable the specific properties you want to bind to, for 
-        // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("Id")] Customer customer)
+        public async Task<IActionResult> Edit(EditUserViewModel model)
         {
-            if (id != customer.Id)
-            {
-                return NotFound();
-            }
-
             if (ModelState.IsValid)
             {
-                try
-                {
-                    _context.Update(customer);
-                    await _context.SaveChangesAsync();
-                }
-                catch (DbUpdateConcurrencyException)
-                {
-                    if (!CustomerExists(customer.Id))
-                    {
-                        return NotFound();
-                    }
-                    else
-                    {
-                        throw;
-                    }
-                }
+                var customer = await _context.Customers
+                    .Include(o => o.User)
+                    .FirstOrDefaultAsync(o => o.Id == model.Id);
+
+                customer.User.Document = model.Document;
+                customer.User.Name = model.Name;
+                customer.User.PhoneNumber = model.CellPhone;
+                customer.User.Address = model.Address;
+
+                await _userHelper.UpdateUserAsync(customer.User);
                 return RedirectToAction(nameof(Index));
             }
-            return View(customer);
+
+            return View(model);
         }
 
         // GET: Customers/Delete/5
@@ -172,26 +170,28 @@ namespace GAPTest.Web.Controllers
             }
 
             var customer = await _context.Customers
-                .FirstOrDefaultAsync(m => m.Id == id);
+                .Include(p => p.User)
+                .Include(p => p.Policies)
+                .FirstOrDefaultAsync(p => p.Id == id);
             if (customer == null)
             {
                 return NotFound();
             }
 
-            return View(customer);
-        }
+            if (customer.Policies.Count > 0)
+            {
+                ModelState.AddModelError(string.Empty, "This customer can't be removed");
+                return RedirectToAction(nameof(Index));
 
-        // POST: Customers/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> DeleteConfirmed(int id)
-        {
-            var customer = await _context.Customers.FindAsync(id);
+
+            }
+
+             await _userHelper.DeleteUserAsync(customer.User.Email);
+
             _context.Customers.Remove(customer);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
-
         private bool CustomerExists(int id)
         {
             return _context.Customers.Any(e => e.Id == id);
